Fix some remove tests.

+++ workingtree.py
@@ -43,6 +43,7 @@ from dulwich.object_store import (
     )
 from dulwich.objects import (
     Blob,
+    Tree,
     S_IFGITLINK,
     )
 from dulwich.repo import Repo
@@ -401,9 +402,6 @@ class GitWorkingTree(workingtree.WorkingTree):
         if len(files) == 0:
             return # nothing to do
 
-        # Sort needed to first handle directory content before the directory
-        files.sort(reverse=True)
-
         def backup(file_to_backup):
             abs_path = self.abspath(file_to_backup)
             backup_name = self.controldir._available_backup_name(file_to_backup)
@@ -411,49 +409,63 @@ class GitWorkingTree(workingtree.WorkingTree):
             return "removed %s (but kept a copy: %s)" % (
                 file_to_backup, backup_name)
 
+        # Sort needed to first handle directory content before the directory
+        files.sort(reverse=True)
+        files_to_backup = []
+
         with self.lock_tree_write():
+            # Bail out if we are going to delete files we shouldn't
+            if not keep_files and not force:
+                for (file_id, path, content_change, versioned, parent_id, name,
+                     kind, executable) in self.iter_changes(self.basis_tree(),
+                         include_unchanged=True, require_versioned=False,
+                         want_unversioned=True, specific_files=files):
+                    if versioned[0] == False:
+                        # The record is unknown or newly added
+                        files_to_backup.append(path[1])
+                    elif (content_change and (kind[1] is not None) and
+                            osutils.is_inside_any(files, path[1])):
+                        # Versioned and changed, but not deleted, and still
+                        # in one of the dirs to be deleted.
+                        files_to_backup.append(path[1])
+
             for f in files:
                 if f == '':
                     continue
-                else:
-                    abs_path = self.abspath(f)
-                    if verbose:
-                        # having removed it, it must be either ignored or unknown
-                        if self.is_ignored(f):
-                            new_status = 'I'
-                        else:
-                            new_status = '?'
-                        # XXX: Really should be a more abstract reporter interface
-                        kind_ch = osutils.kind_marker(self.kind(f))
-                        to_file.write(new_status + '       ' + f + kind_ch + '\n')
-                    # Unversion file
-                    # TODO(jelmer): _unversion_path() is O(size-of-index) for directories
-                    if self._unversion_path(f) == 0:
-                        if (osutils.isdir(abs_path) and
-                            len(os.listdir(abs_path)) == 0):
-                            if not keep_files:
-                                osutils.delete_any(abs_path)
-                            message = "removed %s" % (f,)
-                        else:
-                            message = "%s is not versioned." % (f,)
+
+                try:
+                    kind = self.kind(f)
+                except errors.NoSuchFile:
+                    kind = None
+
+                abs_path = self.abspath(f)
+                if verbose:
+                    # having removed it, it must be either ignored or unknown
+                    if self.is_ignored(f):
+                        new_status = 'I'
                     else:
-                        message = "removed %s" % (f,)
-                        if osutils.lexists(abs_path):
-                            if (osutils.isdir(abs_path) and
-                                len(os.listdir(abs_path)) > 0):
-                                if force:
-                                    osutils.rmtree(abs_path)
-                                    message = "deleted %s" % (f,)
-                                else:
-                                    message = backup(f)
+                        new_status = '?'
+                    # XXX: Really should be a more abstract reporter interface
+                    kind_ch = osutils.kind_marker(kind)
+                    to_file.write(new_status + '       ' + f + kind_ch + '\n')
+                if kind is None:
+                    message = "%s does not exist" % (f, )
+                else:
+                    if not keep_files:
+                        if f in files_to_backup and not force:
+                            backup(f)
+                        else:
+                            if kind == 'directory':
+                                osutils.rmtree(abs_path)
                             else:
-                                if not keep_files:
-                                    osutils.delete_any(abs_path)
-                                    message = "deleted %s" % (f,)
+                                osutils.delete_any(abs_path)
+                    message = "removed %s" % (f,)
+                self._unversion_path(f)
 
                 # print only one message (if any) per file.
                 if message is not None:
                     trace.note(message)
+            self._versioned_dirs = None
             self.flush()
 
     def _add(self, files, ids, kinds):
@@ -1351,13 +1363,17 @@ class InterIndexGitTree(InterGitTrees):
 
     def _iter_git_changes(self, want_unchanged=False, specific_files=None,
             require_versioned=False, include_root=False):
-        # TODO(jelmer): Handle include_root
-        # TODO(jelmer): Handle require_versioned
+        if require_versioned and specific_files:
+            for path in specific_files:
+                if (not self.source.is_versioned(path) and
+                    not self.target.is_versioned(path)):
+                    raise errors.PathsNotVersionedError(path)
         # TODO(jelmer): Restrict to specific_files, for performance reasons.
         with self.lock_read():
             return changes_between_git_tree_and_working_copy(
                 self.source.store, self.source.tree,
-                self.target, want_unchanged=want_unchanged)
+                self.target, want_unchanged=want_unchanged,
+                include_root=include_root)
 
     def compare(self, want_unchanged=False, specific_files=None,
                 extra_trees=None, require_versioned=False, include_root=False,
@@ -1404,14 +1420,17 @@ tree.InterTree.register_optimiser(InterIndexGitTree)
 def untracked_changes(tree):
     for e in tree.extras():
         ap = tree.abspath(e)
-        st = os.stat(ap)
+        st = os.lstat(ap)
         try:
             np, accessible  = osutils.normalized_filename(e)
         except UnicodeDecodeError:
             raise errors.BadFilenameEncoding(
                 e, osutils._fs_enc)
-        yield ((None, np), (None, st.st_mode),
-               (None, blob_from_path_and_stat(ap.encode('utf-8'), st).id))
+        if stat.S_ISDIR(st.st_mode):
+            obj_id = Tree().id
+        else:
+            obj_id = blob_from_path_and_stat(ap.encode('utf-8'), st).id
+        yield ((None, np), (None, st.st_mode), (None, obj_id))
 
 
 def changes_between_git_tree_and_index(store, from_tree_sha, target,
@@ -1425,11 +1444,14 @@ def changes_between_git_tree_and_index(store, from_tree_sha, target,
 
 
 def changes_between_git_tree_and_working_copy(store, from_tree_sha, target,
-        want_unchanged=False, update_index=False):
+        want_unchanged=False, update_index=False, include_root=False):
     """Determine the changes between a git tree and a working tree with index.
 
     """
     blobs = iter_fresh_blobs(target.index, target.abspath('.').encode(sys.getfilesystemencoding()))
     to_tree_sha = commit_tree(store, blobs)
-    return store.tree_changes(from_tree_sha, to_tree_sha, include_trees=True,
-            want_unchanged=want_unchanged, change_type_same=True)
+    for change in store.tree_changes(from_tree_sha, to_tree_sha, include_trees=True,
+            want_unchanged=want_unchanged, change_type_same=True):
+        if change[1][0] == '' and not include_root:
+            continue
+        yield change
